Rename : TemplateComponent.js -> StyledTemplate.js #132
