+++ src/components/common/template/StyledTemplate.js
